Fixed bug in decoding large images

+++ src/libImaging/HexDecode.c
@@ -21,7 +21,7 @@
 		(v >= 'A' && v <= 'F') ? v - 'A' + 10 : -1)
 
 int
-ImagingHexDecode(Imaging im, ImagingCodecState state, UINT8* buf, int bytes)
+ImagingHexDecode(Imaging im, ImagingCodecState state, UINT8* buf, Py_ssize_t bytes)
 {
     UINT8* ptr;
     int a, b;
